requirements.txt
langchain==0.1.17
langchain-openai==0.1.5
langchain_pinecone==0.1.0
python-dotenv==1.0.1
fastapi==0.111.0
streamlit==1.34.0
snowflake_sqlalchemy==1.5.3